Fixed 'Module Unavailable' error in Karma
The karma configuration was using pattern matching to load the files for testing (\scripts\*), but since module.js was being loaded after authenticationService.js, Karma was throwing an error that the ua.security module was not found.  By putting the services into a sub folder (services), the files will be loaded in the correct order.

Also fixed the gruntfile to build the project correctly given the new directory structure.