Fixed sessionStorage mocks to run in all browsers
Load a sessionStorage mock based on the browser under test.  If the browser is PhantomJS, Jasmine's spyOn().andCallFake() is used to mock the sessionStorage.  If the browser is not PhantomJs, a mock object is created and added as a $window property.  This is a workaround for the Firefox not handling spyOn().andCallFake() and PhantomJS defining sessionStorage property as non-configurable (see https://github.com/pivotal/jasmine/issues/299).